Bump to 1.1.28 — align user output schema with format_user() and fix posts_count type
- getUserSchema() now matches the actual format_user() response: drop the
  never-emitted role (singular) and mislabeled avatar_url (key is avatar);
  add missing detailed fields bio, posts_count, last_login, capabilities.
- Cast posts_count to int (count_user_posts() returns a numeric string),
  matching the integer type declared in the schema.

+++ src/Services/UserManager.php
@@ -354,7 +354,7 @@ class UserManager extends AbstractService {
             $data['website'] = $user->user_url;
             $data['bio'] = $user->description;
             $data['avatar'] = get_avatar_url( $user->ID, [ 'size' => 96 ] );
-            $data['posts_count'] = count_user_posts( $user->ID );
+            $data['posts_count'] = (int) count_user_posts( $user->ID );
             $data['last_login'] = get_user_meta( $user->ID, 'last_login', true ) ?: null;
             $data['capabilities'] = array_keys( array_filter( $user->allcaps ) );
         }
